Star Citizen 3.3 OCS Preparation
This patch is being created in the light of the upcoming 3.3 release containing
a new interesting technology called OCS. Everybody is interested in the patch's
performance. Since this is a tool to capture the performance of the game, I have
updated it to bring it up the Reddit's standard. Capturing more information than before,
and automatically generating a Reddit post by the click of a button allowing
everybody to share their setup, performance, and experience.